LayoutTests/imported/w3c/IndexedDB-private-browsing/idbobjectstore_get2-expected.txt


PASS IDBObjectStore.get() - key is a string
